Replaced initial team shuffling algorithm with a much less complex one.
The new algorithm sorts all players by their recent PPH, regardless of team. The player with the lowest PPH is randomly assigned to a team. All remaining players are iterated in pairs from highest to lowest PPH. For each pair, the player with the higher PPH is assigned to the team currently having the lower combined PPH, while the other player goes to the other team.
This may not be as precise as the previous algorithm, but then again PPH isn't an accurate indicator anyway and the vastly reduced complexity is well worth the slightly lower accuracy.